# tests/fixtures/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ClaimRequest": {
        "description": "Запрос claim гостевого заказа после логина.",
        "properties": {
          "guest_capability_token": {
            "maxLength": 512,
            "minLength": 10,
            "title": "Guest Capability Token",
            "type": "string"
          }
        },
        "required": [
          "guest_capability_token"
        ],
        "title": "ClaimRequest",
        "type": "object"
      },
      "ClaimResponse": {
        "description": "Ответ claim гостевого заказа.",
        "properties": {
          "guest_identity": {
            "$ref": "#/components/schemas/GuestIdentity"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "guest_identity",
          "message"
        ],
        "title": "ClaimResponse",
        "type": "object"
      },
...
      "FactoryInfo": {
        "description": "Информация о фабрике.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "FactoryInfo",
        "type": "object"
      },
...
      "GuestIdentity": {
        "properties": {
          "factory_id": {
            "format": "uuid",
            "title": "Factory Id",
            "type": "string"
          },
          "scopes": {
            "items": {
              "$ref": "#/components/schemas/GuestScope"
            },
            "title": "Scopes",
            "type": "array"
          },
          "token_id": {
            "title": "Token Id",
            "type": "string"
          }
        },
        "required": [
          "factory_id",
          "scopes",
          "token_id"
        ],
        "title": "GuestIdentity",
        "type": "object"
      },
      "GuestScope": {
        "description": "Области доступа для гостевого токена.",
        "enum": [
          "read:order",
          "read:bom",
          "*"
        ],
        "title": "GuestScope",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "description": "Запрос на вход (отправка magic link).",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "MessageResponse": {
        "description": "Простой ответ с сообщением.",
        "properties": {
          "dev_magic_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dev Magic Link"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "RegisterRequest": {
        "description": "Запрос на регистрацию фабрики.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "factory_name": {
            "maxLength": 255,
            "minLength": 2,
            "title": "Factory Name",
            "type": "string"
          }
        },
        "required": [
          "email",
          "factory_name"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "description": "Ответ с JWT токеном.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserInfo"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
      "UserInfo": {
        "description": "Информация о пользователе.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "factory": {
            "$ref": "#/components/schemas/FactoryInfo"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_owner": {
            "title": "Is Owner",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "is_owner",
          "factory"
        ],
        "title": "UserInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VerifyRequest": {
        "description": "Проверка magic token.",
        "properties": {
          "token": {
            "maxLength": 64,
            "minLength": 32,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "VerifyRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Furniture AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/claim-guest-order": {
      "post": {
        "description": "Claim гостевого заказа после логина.\n\nВалидация capability-токена и проверка привязки к фабрике:\nТокен должен быть подписан, не истёкшим, не UUID-only.\nФабрика токена должна совпадать с фабрикой залогиненного пользователя.",
        "operationId": "claim_guest_order_api_v1_auth_claim_guest_order_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Claim Guest Order",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Запрос magic link для входа.\n\nВсегда возвращает успех (защита от перебора email).",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Получить информацию о текущем пользователе.",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Регистрация новой фабрики и первого пользователя (owner).\n\nОтправляет magic link на email для подтверждения.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/verify": {
      "post": {
        "description": "Проверка magic token и выдача JWT.\n\nToken можно использовать только один раз.",
        "operationId": "verify_api_v1_auth_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}